// front/src/ts/enums/enums.ts
// Alert related enums
export enum AlertTypeEnum {
  Error = 'error',
  Warning = 'warning',
  Success = 'success'
}

// Notes related enums
export enum NotesElementTypeEnum {
  Weak = 'weak',
  Medium = 'medium',
  Strong = 'strong'
}

export enum NotesBtnDirectionEnum {
  Right = 'right',
  Left = 'left',
}

// Interaction related enums
export enum InteractionTypeEnum {
  Expression = 'expression',
  Word = 'word'
}